refactor: rename & change folder structure
