+++ src/modules/ama/new-ama/broadcast-ama.ts
@@ -2,7 +2,7 @@ import { Context } from "telegraf";
 import { UUID_PATTERN, validateIdPattern } from "../helper/utils";
 import { CALLBACK_ACTIONS } from "../ama.constants";
 import { AMA, PublicGroupInfo } from "../types";
-import { buildAMAMessage, imageUrl } from "../helper/msg-builder";
+import { buildAMAMessage, imageUrl } from "./helper/msg-builder";
 import { UUID } from "crypto";
 
 export async function handleBroadcastNow(
